# internal/docs/openapi.yaml
openapi: 3.0.3
info:
  title: fusion-bff API
  description: |
    Backend for Frontend for the fusion platform GUI. Validates OIDC sessions, enforces RBAC, and proxies authenticated requests to fusion-forge, fusion-index, fusion-weave, fusion-wizard, and fusion-content.
  version: 0.12.0
  license:
    name: GPL-3.0
servers:
- url: /
...
paths:
...
  /bff/admin/owner-groups:
    get:
      tags:
      - rbac-admin
      summary: List owner groups
      security:
      - sessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OwnerGroupRow'
    post:
      tags:
      - rbac-admin
      summary: Create an owner group (a team owning CRs)
      security:
      - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  description: 1-63 chars, letters/digits/./_/-, alphanumeric at both ends
                description:
                  type: string
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerGroupRow'
        '400':
          description: Invalid body or unknown owner group
        '409':
          description: Already exists
  /bff/admin/owner-groups/{id}:
    delete:
      tags:
      - rbac-admin
      summary: Delete owner group
      security:
      - sessionCookie: []
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        '204':
          description: Deleted
        '400':
          description: Invalid id
        '404':
          description: Not found
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sid
      description: HttpOnly session cookie set by /bff/callback
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC access token, accepted as a fallback when no session cookie is present
  schemas:
...
    OwnerGroupRow:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
...